+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Achievements added in 11.0.7
 - Feature to prioritize which tab is used when getting redirected to an achievement
+- Has Reward filter (dev note: reward type filters will come in a later version)
+- Option to show a Warband icon on Warband achievements
+- Option to hide the blue Warband achievement color
 
 ## 79.0 - 2024-11-15
 ### Added